[NFC] Move DroppedVariableStats to its own file
